refactor(ExperimentalUtils): backport get_image_data
